refactor(lib): consolidate post collection utils
