+++ src/pages/index.astro
@@ -4,7 +4,7 @@ import PostPreview from '@/components/blog/PostPreview';
 import SocialList from '@/components/SocialList';
 import Heading from '@/components/Heading';
 import Paragraph from '@/components/Paragraph';
-import { sortMDByDate } from '@/utils';
+import { sortMDByDate } from '@/lib/post';
 import { getPostsCollection } from '@/lib/post';
 
 const MAX_POSTS = 10;
