+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export { getFormattedDate } from './date';
 export { elementHasClass, toggleClass } from './domElement';
-export { sortMDByDate, getUniqueTags, getUniqueTagsWithCount } from './post';
